requirements.txt
nvidia-ml-py==12.570.86
PyYAML==6.0.2
